fix: namespace identifier type
