+++ app/graphql/types/path_type.rb
@@ -2,6 +2,6 @@ module Types
   class PathType < Types::BaseObject
     field :elevation, [Types::JpElevationType, null: true], null: true
     field :lineString, String, null: true
-    field :stopPoints, [Types::IdentifierType, null: true], null: true
+    field :stopPoints, [Types::Tfl::Entities::IdentifierType, null: true], null: true
   end
 end
