+++ app/graphql/types/tfl/entities/identifier_type.rb
@@ -1,0 +1,16 @@
+module Types
+  module Tfl
+    module Entities
+      class IdentifierType < Types::BaseObject
+        field :crowding, Types::Tfl::Entities::CrowdingType, null: true
+        field :fullName, String, null: true
+        field :id, String, null: true
+        field :name, String, null: true
+        field :routeType, Types::RouteType, null: true
+        field :status, Types::StatusType, null: true
+        field :type, String, null: true
+        field :uri, String, null: true
+      end
+    end
+  end
+end
