+++ app/graphql/types/tfl/entities/stop_point_type.rb
@@ -16,7 +16,7 @@ module Types
         field :lat, Float, null: true
         field :lineGroup, [Types::Tfl::Entities::LineGroupType, null: true], null: true
         field :lineModeGroups, [Types::Tfl::Entities::LineModeGroupType, null: true], null: true
-        field :lines, [Types::IdentifierType, null: true], null: true
+        field :lines, [Types::Tfl::Entities::IdentifierType, null: true], null: true
         field :lon, Float, null: true
         field :modes, [String, null: true], null: true
         field :naptanId, String, null: true
